Refactor (article-category) : 상수 관리 방식 변경 (object → const assertion)
