+++ app/api/database/category/route.ts
@@ -1,7 +1,7 @@
 import { NextResponse } from "next/server";
 import { getHeaders, notionDatabaseUrl } from "@/app/lib/data/notion";
 import { DatabaseObjectResponse } from "@notionhq/client/build/src/api-endpoints";
-import { CategoryResponse } from "@/app/types/notion";
+import { ArticleCategory, CategoryResponse } from "@/app/types/notion";
 
 export const GET = async () => {
     try {
@@ -18,7 +18,7 @@ export const GET = async () => {
 
         if (categoryProperty.type !== "select") throw new Error("Wrong Access");
 
-        const categories: CategoryResponse = categoryProperty.select.options.map(({ name }) => ({ category: name }));
+        const categories: CategoryResponse = categoryProperty.select.options.map(({ name }) => ({ category: name as keyof typeof ArticleCategory }));
         return NextResponse.json({ categories: categories });
     } catch (error) {
         const message = error instanceof Error ? error.message : String(error);
