+++ app/components/articles/CategoryButton.tsx
@@ -1,11 +1,11 @@
 'use client'
 
-import { ArticleCategory } from "@/app/lib/data/notion";
+import { ArticleCategory } from "@/app/types/notion";
 import { CSSProperties, MouseEvent } from "react";
 import { useRouter } from "next/navigation";
 import Button from "../Button";
 
-export default function CategoryButton({ category, style }: { category: string; style?: CSSProperties }) {
+export default function CategoryButton({ category, style }: { category: keyof typeof ArticleCategory; style?: CSSProperties }) {
     const router = useRouter();
 
     const handleCategoryClick = (e: MouseEvent) => {
