+++ app/lib/data/navItem.ts
@@ -1,4 +1,4 @@
-import { ArticleCategory } from "./notion";
+import { ArticleCategory } from "@/app/types/notion";
 
 interface NavItem {
     path: string;
@@ -7,6 +7,6 @@ interface NavItem {
 
 export const navItem: NavItem[] = [
     { path: "/", title: "홈" },
-    ...Object.keys(ArticleCategory).map((category) => ({ path: `/posts/${category}`, title: ArticleCategory[category] })),
+    ...Object.keys(ArticleCategory).map((category: keyof typeof ArticleCategory) => ({ path: `/posts/${category}`, title: ArticleCategory[category] })),
     { path: "/projects", title: "프로젝트" }
 ];
