+++ app/posts/[category]/layout.tsx
@@ -1,9 +1,7 @@
 import { baseUrl } from "@/app/lib/data/api";
-import { ArticleCategory } from "@/app/lib/data/notion";
-import { CategoryResponse } from "@/app/types/notion";
+import { ArticleCategory, CategoryResponse } from "@/app/types/notion";
 import PageHeader from "@/app/components/PageHeader";
 import Image from "next/image";
-import { getTheme } from "@/app/lib/functions/theme";
 
 export const generateStaticParams = async () => {
     const response = await fetch(`${baseUrl}/api/database/category`);
@@ -12,7 +10,7 @@ export const generateStaticParams = async () => {
     return categories;
 };
 
-export default function CategoryLayout({ params, children }: { params: { category: string }; children: React.ReactNode }) {
+export default function CategoryLayout({ params, children }: { params: { category: keyof typeof ArticleCategory }; children: React.ReactNode }) {
     return (
         <main className="main-wrapper">
             <PageHeader title={ArticleCategory[params.category]}>
